+++ collective/folderishtraverse/browser/traverse_view.py
@@ -8,6 +8,11 @@ from Products.CMFPlone.interfaces.siteroot import IPloneSiteRoot
 from Products.Five.browser import BrowserView
 from Products.statusmessages.interfaces import IStatusMessage
 from plone.folder.interfaces import IFolder
+try:
+    from Products.LinguaPlone.interfaces import ITranslatable
+    LINGUA_PLONE_INSTALLED = True
+except ImportError:
+    LINGUA_PLONE_INSTALLED = False
 
 
 _ = MessageFactory('collective.folderishtraverse')
@@ -37,32 +42,42 @@ class TraverseView(BrowserView):
         url = None
 
         if self.anonymous:
-            def find_endpoint(obj):
-                if not IFolder.providedBy(obj) and\
-                   not IPloneSiteRoot.providedBy(obj): return obj
+            def find_endpoint(obj, lang):
+                if not IFolder.providedBy(obj) \
+                  and not IPloneSiteRoot.providedBy(obj):
+                    return obj
                 contents = obj.contentIds()
                 # TODO: make list reversable
                 for id in contents:
                     child = obj[id]
-
+                    # try to get translation if LinguaPlone is installed, child
+                    # is translatable and child language does not match lang
+                    if LINGUA_PLONE_INSTALLED:
+                        if ITranslatable.providedBy(child)\
+                          and child.Language() != lang:
+                            translation = child.getTranslation(lang)
+                            if not translation:
+                                break
+                            child = translation
                     # only traverse to published objects
                     try:
                         state = wft.getInfoFor(child, 'review_state')
                     except:
                         state = None
-                    if not state == 'published': continue
-
+                    if not state == 'published':
+                        continue
                     # only traverse to objects listed in typesToList
-                    if child.portal_type not in types: continue
-
+                    if child.portal_type not in types:
+                        continue
                     # we've found a published object, which can be used as
                     # possible endpoint, except it has 'traverse_view' enabled
                     obj = child
                     if child.defaultView() == 'traverse_view':
-                        obj = find_endpoint(child)
+                        obj = find_endpoint(child, lang)
                     break
                 return obj
-            ctx = find_endpoint(ctx)
+            plt = getToolByName(self.context, 'portal_languages')
+            ctx = find_endpoint(ctx, plt.getPreferredLanguage())
 
         url = ctx.absolute_url()
         if ctx.defaultView() == 'traverse_view':